--- AA.java
@@ -1,25 +1,0 @@
-// Person class
-
-public class AA {
-    private String firstName;
-    private String lastName;
-    private String contactNumber;
-
-    public AA(String firstName, String lastName, String contactNumber) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.contactNumber = contactNumber;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getContactNumber() {
-        return contactNumber;
-    }
-}
