--- AD.java
@@ -1,6 +1,0 @@
-// Vehicle class
-
-public abstract class AD {
-    public abstract void displayVariants();
-    public abstract double getPrice(int choice);
-}
