+++ Person.java
@@ -1,0 +1,25 @@
+// Person class
+
+public class Person {
+    private String firstName;
+    private String lastName;
+    private String contactNumber;
+
+    public Person(String firstName, String lastName, String contactNumber) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.contactNumber = contactNumber;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public String getContactNumber() {
+        return contactNumber;
+    }
+}
